# Optimize bytestream.Write by double buffering in distributed client