@vanja-p 2025-09-22
I tested this in dev, and it's about 10-20% faster for writing blobs which are 20MB uncompressed, and 6.3MB compressed.